+++ tests/ui/specs/Slider/Sliders-specs.js
@@ -12,17 +12,17 @@ describe('Slider', function () {
 
 			describe('5-way', function () {
 				it('should increment the value of horizontal slider on arrow right key when active', async function () {
-					expect(await slider.self.isFocused()).to.be.true();
+					expect(await slider.self.isFocused()).toBe(true);
 					const originalValue = await slider.knobPositionHorizontal();
 					await Page.spotlightSelect();
 					await Page.spotlightRight();
 					// expect knob `left` css prop to be bigger than original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.be.above(originalValue);
+					expect(newValue).toBeGreaterThan(originalValue);
 				});
 
 				it('should decrement the value of horizontal slider on arrow left key when active', async function () {
-					expect(await slider.self.isFocused()).to.be.true();
+					expect(await slider.self.isFocused()).toBe(true);
 					await Page.spotlightSelect();
 					await Page.spotlightRight();
 					await Page.spotlightRight();
@@ -30,18 +30,18 @@ describe('Slider', function () {
 					await Page.spotlightLeft();
 					// expect knob `left` css prop to be smaller than original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.be.below(originalValue);
+					expect(newValue).toBeLessThan(originalValue);
 				});
 			});
 
 			describe('pointer', function () {
 				it('should change the value of horizontal slider on slider click at position', async function () {
-					expect(await slider.self.isFocused()).to.be.true();
+					expect(await slider.self.isFocused()).toBe(true);
 					const originalValue = await slider.knobPositionHorizontal();
 					await slider.self.click();
 					// expect knob `left` css prop to be bigger than original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.be.above(originalValue);
+					expect(newValue).toBeGreaterThan(originalValue);
 				});
 			});
 		});
@@ -56,7 +56,7 @@ describe('Slider', function () {
 					await Page.spotlightRight();
 					// expect knob `left` css prop to be bigger than original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 
 				it('should not activate without select and cannot decrement the value', async function () {
@@ -67,7 +67,7 @@ describe('Slider', function () {
 					await Page.spotlightLeft();
 					// expect knob `left` css prop to be smaller than original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 				it('should activate on select and can increment the value', async function () {
 					await slider.focus();
@@ -76,7 +76,7 @@ describe('Slider', function () {
 					await Page.spotlightRight();
 					// expect knob `left` css prop to be bigger than original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.be.above(originalValue);
+					expect(newValue).toBeGreaterThan(originalValue);
 				});
 
 				it('should activate on select and can decrement the value', async function () {
@@ -88,7 +88,7 @@ describe('Slider', function () {
 					await Page.spotlightLeft();
 					// expect knob `left` css prop to be smaller than original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.be.below(originalValue);
+					expect(newValue).toBeLessThan(originalValue);
 				});
 			});
 		});
@@ -98,7 +98,7 @@ describe('Slider', function () {
 
 			it('fill bar should be greater than 0 when value is at minimum', async function () {
 				await slider.focus();
-				expect(await slider.sliderFillWidth()).to.be.above(0);
+				expect(await slider.sliderFillWidth()).toBeGreaterThan(0);
 			});
 		});
 
@@ -114,7 +114,7 @@ describe('Slider', function () {
 					await Page.spotlightUp();
 					// expect knob `bottom` css prop to be greater than original one
 					const newValue = await slider.knobPositionVertical();
-					expect(newValue).to.be.above(originalValue);
+					expect(newValue).toBeGreaterThan(originalValue);
 				});
 
 				it('should decrement the value of vertical slider on arrow down key when active', async function () {
@@ -125,7 +125,7 @@ describe('Slider', function () {
 					await Page.spotlightDown();
 					// expect knob `bottom` css prop to be less than original one
 					const newValue = await slider.knobPositionVertical();
-					expect(newValue).to.be.below(originalValue);
+					expect(newValue).toBeLessThan(originalValue);
 				});
 			});
 
@@ -136,7 +136,7 @@ describe('Slider', function () {
 					await slider.self.click();
 					// expect knob `left` css prop to be greater than original one
 					const newValue = await slider.knobPositionVertical();
-					expect(newValue).to.be.above(originalValue);
+					expect(newValue).toBeGreaterThan(originalValue);
 				});
 			});
 		});
@@ -152,7 +152,7 @@ describe('Slider', function () {
 					await Page.spotlightRight();
 					// expect knob `left` css prop to be equal to original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 
 				it('should not decrement the value of horizontal disabled slider on arrow left key when active', async function () {
@@ -164,7 +164,7 @@ describe('Slider', function () {
 					await Page.spotlightLeft();
 					// expect knob `left` css prop to be equal to original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 			});
 
@@ -176,7 +176,7 @@ describe('Slider', function () {
 					await slider.self.click();
 					// expect knob `left` css prop to be equal to original one
 					const newValue = await slider.knobPositionHorizontal();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 			});
 		});
@@ -193,7 +193,7 @@ describe('Slider', function () {
 					await Page.spotlightUp();
 					// expect knob `bottom` css prop to be equal to original one
 					const newValue = await slider.knobPositionVertical();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 
 				it('should not decrement the value of vertical disabled slider on arrow up key when active', async function () {
@@ -204,7 +204,7 @@ describe('Slider', function () {
 					await Page.spotlightDown();
 					// expect knob `bottom` css prop to be equal to original one
 					const newValue = await slider.knobPositionVertical();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 			});
 
@@ -216,7 +216,7 @@ describe('Slider', function () {
 					await slider.self.click();
 					// expect knob `left` css prop to be equal to original one
 					const newValue = await slider.knobPositionVertical();
-					expect(newValue).to.equal(originalValue);
+					expect(newValue).toBe(originalValue);
 				});
 			});
 		});