+++ useScroll/Scrollbar.js
@@ -88,7 +88,8 @@ const useThemeScrollbar = (props) => {
  * @ui
  * @private
  */
-const ScrollbarBase = memo((props) => {
+const ScrollbarBase = memo(({css = componentCss, minThumbSize = 120, vertical = true, ...rest}) => {
+	const props = {css, minThumbSize, vertical, ...rest};
 	const {
 		restProps,
 		scrollbarProps,
@@ -149,12 +150,6 @@ ScrollbarBase.propTypes = /** @lends sandstone/useScroll.Scrollbar.prototype */ 
 	 * @public
 	 */
 	vertical: PropTypes.bool
-};
-
-ScrollbarBase.defaultProps = {
-	css: componentCss,
-	minThumbSize: 120,
-	vertical: true
 };
 
 /**